# optimize boot time pt. 2