This includes:
- removed debug capabilities from kernel
- removed unneeded kernel modules
- removed network service if it's a raspberry 1.3
- removed utmp systemd services

We're down to:
Startup finished in 784ms (kernel) + 5.706s (userspace) = 6.491s 
multi-user.target reached after 5.677s in userspace

@htruong would you mind testing this with the Raspberry Zero W to check if wifi still works?