@dgsiegel 2020-11-13
Together with a few final tweaks, like `initial_turbo` we're now under 5 seconds

Startup finished in 657ms (kernel) + 4.079s (userspace) = 4.736s 
multi-user.target reached after 4.063s in userspace