@htruong 2020-11-19
Works for me:

```
Startup finished in 758ms (kernel) + 5.562s (userspace) = 6.321s
multi-user.target reached after 5.501s in userspace
```